Adding block_t typedef to BlockId, to ensure BlockId does not get negative values.  Also changed the type for stored number of blocks in various classes to size_t.

+++ source/SAMRAI/hier/Box.C
@@ -721,7 +721,7 @@ void
 Box::putToIntBuffer(
    int* buffer) const
 {
-   buffer[0] = d_block_id.getBlockValue();
+   buffer[0] = static_cast<int>(d_block_id.getBlockValue());
    ++buffer;
 
    d_id.putToIntBuffer(buffer);
