+++ source/SAMRAI/hier/BoxNeighborhoodCollection.C
@@ -712,7 +712,8 @@ BoxNeighborhoodCollection::putToRestart(
                const Box& nbr = *ni;
                local_ids[counter] = nbr.getLocalId().getValue();
                ranks[counter] = nbr.getOwnerRank();
-               block_ids[counter] = nbr.getBlockId().getBlockValue();
+               block_ids[counter] =
+                  static_cast<int>(nbr.getBlockId().getBlockValue());
                periodic_ids[counter] = nbr.getPeriodicId().getPeriodicValue();
                db_box_array[counter] = nbr;
                ++counter;
